+++ src/entities/matrix_element.py
@@ -1,6 +1,0 @@
-class MatrixElement:
-    def __init__(self,action=None, _type="empty",tier=1, points=0):
-        self.type=_type
-        self.points=points
-        self.action=action
-        self.tier=tier