+++ src/index.py
@@ -1,7 +1,7 @@
 import sqlite3
 from ui.gui import View
 from services.game import Game
-from entities.snake import Snake
+from services.snake import Snake
 from repositories.score import Score
 db = sqlite3.connect("src/database.db")
 score=Score(db)