+++ src/services/game.py
@@ -1,10 +1,12 @@
 from random import choice, randint
-from entities.matrix_element import MatrixElement
 from entities.default_treat import DefaultTreat
-from entities.purge_treat import PurgeTreat
-empty = MatrixElement()
-treat_1=MatrixElement(DefaultTreat(3), "treat", 1, 1)
-snake_body=MatrixElement(_type="snake")
+from entities.custom_matrix_element import CustomMatrixElement
+from services.treat_factory import TreatFactory
+treat_factory=TreatFactory()
+treat_1=DefaultTreat(3)
+snake_body=CustomMatrixElement("snake")
+empty=CustomMatrixElement("empty")
+
 START=[
 [empty, empty, empty, empty, empty, empty, empty,
 empty, empty, empty, empty, empty, empty, empty],
@@ -35,13 +37,24 @@ GAME_OVER=[[0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0],
         [0, 2, 2, 0, 0, 2, 0, 2, 2, 2, 0, 2, 2, 0]]
 
 class Game:
+    """Game is class responsible for integrating entities, and producing the image
+    supplied for gui."""
     def __init__(self, snake, score, game_matrix=START):
+        """self.points is the current amount of points
+        the player has collected."""
         self.points=0
+        """self.score is an instance of the Score class."""
         self.score=score
         self.player_name=""
+        """self.start_position is the initial position
+        of the game. This never changes."""
         self.start_position=[x[:] for x in game_matrix]
+        """self.game_matrix is the current position of the game."""
         self.game_matrix=[x[:] for x in game_matrix]
+        """self.snake is an instance of the snake-class."""
         self.snake=snake
+        """self.direction is the direction given to the snake
+        instance on advance."""
         self.direction=1
         self.game_over=True
         for i in snake.position:
@@ -52,6 +65,7 @@ class Game:
                 self.coordinates.append([i, j])
 
     def start(self, name):
+        """this method resets the game"""
         self.points=0
         self.game_matrix=[x[:] for x in self.start_position]
         self.player_name=name
@@ -60,25 +74,33 @@ class Game:
         self.game_over=False
 
     def set_game_matrix(self, matrix):
+        """This method allows high customization of special treat functionality
+        in the future."""
         self.game_matrix=matrix
 
     def change_direction(self, direction):
         self.direction=direction
 
     def out_out_bounds(self, head):
+        """This method tests whether the snake has hit wall"""
         return (head[1]<0 or head[0]>=len(self.game_matrix) or
         head[0]<0 or head[1]>=len(self.game_matrix[0]))
 
     def square_is_free(self, coordinates):
+        """This tests whether the square that snake's head is currently on its body.
+        Returns true if snake blocks do not overlap and false otherwise."""
         return self.game_matrix[coordinates[0]][coordinates[1]].type!="snake"
 
     def purge_candy(self):
+        """This method removes all candies from the map. Used by the PurgeTreat class's
+        consume method."""
         for i in self.game_matrix:
             for j, elem in enumerate(i):
                 if elem.type!="snake":
                     i[j]=empty
 
     def clear_game_matrix(self):
+        """This clears previous snake blocks. Used before rendering new snake position."""
         for i in self.game_matrix:
             for j, elem in enumerate(i):
                 if elem.type=="snake":
@@ -91,53 +113,53 @@ class Game:
         return self.game_matrix[coordinate[0]][coordinate[1]].type!="snake"
 
     def get_non_snake_coordinates(self):
+        """This method is used to find squares on the game_matrix that are not snake body."""
         non_snake_coordinates=list(filter(self.no_snake_filter, self.coordinates))
         return non_snake_coordinates
 
     def get_empty_coordinates(self):
+        """This is used to find free squares systematically to reduce overhead caused
+        by random polling"""
         free_coordinates = list(filter(self.free_filter, self.coordinates))
         return free_coordinates
 
-    def generate_default_treat(self):
-        free_coordinates=self.get_empty_coordinates()
+    def new_treat(self):
+        new_treat=TreatFactory().new_random_treat()
+        free_coordinates=self.get_non_snake_coordinates()
+        if new_treat.tier==1:
+            free_coordinates=self.get_empty_coordinates()
         if len(free_coordinates)==0:
             return
-        new_effect=choice([-1, -2, 1, 1, 1, 1, 1, 1, 1])
-        treat_coordinates = choice(free_coordinates)
-        x_coordinate=treat_coordinates[0]
-        y_coordinate=treat_coordinates[1]
-        new_element=MatrixElement(DefaultTreat(new_effect), "treat", 1, 1)
-        self.game_matrix[x_coordinate][y_coordinate]=new_element
-
-    def generate_matrix_treat(self):
-        coordinates=choice(self.get_non_snake_coordinates())
-        new_element=MatrixElement(PurgeTreat("?"), "matrix_treat", 5, 20)
-        self.game_matrix[coordinates[0]][coordinates[1]]=new_element
-
-    def new_treat(self):
-        lottery = randint(1, 90)
-        if lottery==90:
-            self.generate_matrix_treat()
-            return
-        self.generate_default_treat()
+        coordinates=choice(free_coordinates)
+        self.game_matrix[coordinates[0]][coordinates[1]]=new_treat
 
     def eat_treat(self, treat):
+        """This method is called when snake's head is on any consumable element.
+        It assures the consume function of the element is called with a correct argument."""
         self.points+=treat.points
         if treat.type=="treat":
-            treat.action.consume(self.snake)
+            treat.consume(self.snake)
         if treat.type=="matrix_treat":
-            treat.action.consume(self)
+            treat.consume(self)
+        if treat.type=="dual_treat":
+            print("ny consumee")
+            treat.consume(self, self.snake)
 
     def is_treat(self, head):
-        if self.game_matrix[head[0]][head[1]].type in ["treat", "matrix_treat"]:
+        """This method tests whether the snake's head is currently on a consumable element.
+        Returns true if yes, false otherwise."""
+        if self.game_matrix[head[0]][head[1]].tier!=0:
             return True
         return False
 
     def update_game_matrix(self,snake):
+        """This method renders snake's body on the game_matrix."""
         for block in snake:
             self.game_matrix[block[0]][block[1]]=snake_body
 
     def advance(self):
+        """This is the game's tick method. It manages everything that is needed to produce
+        the next image supplied to GUI, and then returns it."""
         snake_image = self.snake.advance(self.direction)
         head = snake_image[len(snake_image)-1]
         if not self.game_over and (self.out_out_bounds(head) or not self.square_is_free(head)):