+++ src/tests/flood_treat_test.py
@@ -1,0 +1,7 @@
+from entities.flood_treat import FloodTreat
+import unittest
+class TestFloodTreat(unittest.TestCase):
+    def testUp(self):
+        pass
+    def test_consume(self):
+        pass